Hide/Show NavButtons based on prescence of user_token.
