+++ src/App.js
@@ -1,22 +1,28 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import './App.css';
-import NavButton from './components/navbutton.js';
-import Shopping from './components/shopping.js';
-import AddItem from './components/addItem.js';
+import { USER_TOKEN } from './constants';
+import getToken from './tokenScript';
+import NavButton from './components/navbutton';
+import Shopping from './components/shopping';
+import AddItem from './components/addItem';
 import { useToken } from './lib/useToken';
+import './App.css';
 
 function App() {
-  const [userToken, refreshToken] = useToken();
-  console.log(`user token: ${userToken}`);
+  const [userToken, setToken] = useToken();
+
+  function handleClick() {
+    const token = getToken();
+    localStorage.setItem(USER_TOKEN, token);
+    setToken(token);
+  }
+
   if (userToken) {
-    refreshToken();
     return (
       <Router>
         <div className="App">
           <Route exact path="/" component={Shopping} />
           <Route path="/add" component={AddItem} />
-
           <nav id="nav">
             <NavButton path="/" text="Shopping" />
             <NavButton path="/add" text="Add Item" />
@@ -26,12 +32,11 @@ function App() {
     );
   } else {
     return (
-      <Router>
-        <div className="App">
-          <Route exact path="/" component={Shopping} />
-          <Route path="/add" component={AddItem} />
-        </div>
-      </Router>
+      <div className="App">
+        <button className="button" onClick={handleClick}>
+          Create New List
+        </button>
+      </div>
     );
   }
 }
